refactor: Package 이름 수정 entity -> config
