+++ src/main/java/com/baro/oauth/infra/kakao/KakaoOAuthClient.java
@@ -4,9 +4,9 @@ import com.baro.oauth.application.OAuthClient;
 import com.baro.oauth.application.dto.OAuthMemberInfo;
 import com.baro.oauth.application.dto.OAuthTokenInfo;
 import com.baro.oauth.domain.OAuthServiceType;
+import com.baro.oauth.infra.kakao.config.KakaoOAuthProperty;
 import com.baro.oauth.infra.kakao.dto.KakaoMemberResponse;
 import com.baro.oauth.infra.kakao.dto.KakaoTokenResponse;
-import com.baro.oauth.infra.kakao.entity.KakaoOAuthProperty;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
